requirements.txt
websockets==14.2
fastapi[standard]==0.115.8
uvicorn[standard]==0.34.0
apscheduler==3.11.0
requests==2.32.3
python-socketio==5.12.1
aiohttp==3.11.12